requirements.txt
absl-py==0.11.0
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.6
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
iniconfig==1.1.1
Markdown==3.3.3
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.1.2
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tdqm==0.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
